\name{PopMCarmatest-package}
\alias{PopMCarmatest-package}
\alias{PopMCarmatest}
\docType{package}
\title{
\packageTitle{PopMCarmatest}
}
\description{
\packageDescription{PopMCarmatest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PopMCarmatest}
\packageIndices{PopMCarmatest}
An overview of how to use the package, including the most important functions
}
\author{Juho Kopra
\packageAuthor{PopMCarmatest}

Maintainer: Juho Kopra\packageMaintainer{PopMCarmatest}
}
\references{
Literature or other references for background information
}
Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.

}
\examples{
simple examples of the most important functions
}
